discord-bots/pause-calendar/requirements.txt
discord.py==2.4.0
APScheduler==3.10.4
pytz==2024.1
python-dateutil==2.9.0.post0
